test: Add `utils.to_codepoints` to help debug failed tests
This utility function in the `spec` module creates a human-readable
array of codepoints from a string. Instead of `assert.are.equal` with
the expected and the actual string, we can use `assert.are.same` with
the expected and the actual array of codepoints, and failed tests will
pinpoint the exact codepoint that is different.